# docs/02-api-specification.yaml
openapi: 3.0.3
info:
  title: CMS kéo-thả API
  version: 0.2.0
  description: >
    API cho Editor Engine, Content Editor UI và Renderer.
    Mọi endpoint yêu cầu header Authorization: Bearer <token> trừ khi ghi chú khác.
    Phiên bản 0.2.0 — Cập nhật sau SA Review 17/08/2026:
    Bổ sung pagination, refresh token, DELETE/PATCH, siteId trên content-items, Error schema chuẩn, Sites_Users management, media upload per-site, X-Request-ID.
servers:
  - url: https://api.example.com/v1
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Đăng nhập, trả về access + refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:    { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Đăng nhập thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:  { type: string, description: "JWT, TTL 15 phút" }
                  refresh_token: { type: string, description: "Opaque token, TTL 30 ngày" }
                  expires_in:    { type: integer, example: 900, description: "seconds" }
                  user:          { $ref: '#/components/schemas/User' }
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Làm mới access token bằng refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        '200':
          description: Token mới
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  expires_in:   { type: integer }
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /sites/{siteId}/pages:
    get:
      tags: [Pages]
      summary: Danh sách trang trong 1 site
      parameters:
        - $ref: '#/components/parameters/SiteId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: status
          in: query
          schema: { type: string, enum: [draft, published] }
      responses:
        '200':
          description: Danh sách page (paginated)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PagedPages' }
    post:
      tags: [Pages]
      summary: Tạo page mới (chỉ role designer/owner)
      parameters:
        - $ref: '#/components/parameters/SiteId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slug]
              properties:
                slug:     { type: string }
                seo_meta: { type: object, additionalProperties: true }
      responses:
        '201':
          description: Đã tạo
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Page' }
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /sites/{siteId}/content-types:
    get:
      tags: [ContentTypes]
      summary: Danh sách content type đã định nghĩa
      parameters:
        - $ref: '#/components/parameters/SiteId'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ContentType' }
    post:
      tags: [ContentTypes]
      summary: Tạo content type mới (chỉ designer/owner)
      parameters:
        - $ref: '#/components/parameters/SiteId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContentTypeInput' }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContentType' }
  /sites/{siteId}/content-types/{contentTypeId}:
    patch:
      tags: [ContentTypes]
      summary: Cập nhật content type (chỉ designer/owner)
      parameters:
        - $ref: '#/components/parameters/SiteId'
        - $ref: '#/components/parameters/ContentTypeId'
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContentTypeInput' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContentType' }
    delete:
      tags: [ContentTypes]
      summary: Xóa content type (chỉ owner)
      parameters:
        - $ref: '#/components/parameters/SiteId'
        - $ref: '#/components/parameters/ContentTypeId'
      responses:
        '204':
          description: Đã xóa

  # ─── CONTENT ITEMS ───────────────────────────────────────────
  /sites/{siteId}/content-types/{contentTypeId}/items:
    get:
      tags: [ContentItems]
      summary: Danh sách content item (paginated)
      parameters:
        - $ref: '#/components/parameters/SiteId'
        - $ref: '#/components/parameters/ContentTypeId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: status
          in: query
          schema: { type: string, enum: [draft, published] }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PagedContentItems' }
    post:
      tags: [ContentItems]
      summary: Tạo content item mới (editor trở lên)
      parameters:
        - $ref: '#/components/parameters/SiteId'
        - $ref: '#/components/parameters/ContentTypeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fields: { type: object, additionalProperties: true }
                status: { type: string, enum: [draft, published] }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContentItem' }
  /sites/{siteId}/content-types/{contentTypeId}/items/{itemId}:
    get:
      tags: [ContentItems]
      summary: Lấy chi tiết 1 content item
      parameters:
        - $ref: '#/components/parameters/SiteId'
        - $ref: '#/components/parameters/ContentTypeId'
        - $ref: '#/components/parameters/ItemId'
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContentItem' }
    patch:
      tags: [ContentItems]
      summary: Cập nhật content item
      parameters:
        - $ref: '#/components/parameters/SiteId'
        - $ref: '#/components/parameters/ContentTypeId'
        - $ref: '#/components/parameters/ItemId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                fields: { type: object, additionalProperties: true }
                status: { type: string, enum: [draft, published] }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContentItem' }
    delete:
      tags: [ContentItems]
      summary: Xóa mềm content item
      parameters:
        - $ref: '#/components/parameters/SiteId'
        - $ref: '#/components/parameters/ContentTypeId'
        - $ref: '#/components/parameters/ItemId'
      responses:
        '204':
          description: Đã xóa (soft delete)

  # ─── MEDIA ───────────────────────────────────────────────────
...
  /ai/suggest/seo-meta:
    post:
      tags: [AI]
      summary: Gợi ý SEO meta description từ nội dung trang (dùng Gemini Flash)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [page_content]
              properties:
                page_content: { type: string, description: "Text content của trang, max 4000 chars" }
                language:     { type: string, default: "vi" }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta_description: { type: string, maxLength: 160 }
                  title_suggestion: { type: string, maxLength: 60 }
  /ai/suggest/excerpt:
    post:
      tags: [AI]
      summary: Gợi ý excerpt cho content item (blog, sản phẩm)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body:     { type: string }
                language: { type: string, default: "vi" }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  excerpt: { type: string, maxLength: 300 }
components:
  parameters:
    SiteId:
      name: siteId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    ContentTypeId:
      name: contentTypeId
      in: path
      required: true
      schema: { type: string, format: uuid }
    ItemId:
      name: itemId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
    Cursor:
      name: cursor
      in: query
      description: "Cursor-based pagination — giá trị từ trường `next_cursor` của response trước"
      schema: { type: string }
  responses:
    Unauthorized:
      description: Chưa đăng nhập hoặc token hết hạn
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Không đủ quyền
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:       { type: string, example: "RESOURCE_NOT_FOUND", description: "Machine-readable error code" }
        message:    { type: string, example: "Page không tồn tại", description: "Human-readable message" }
        details:    { type: object, additionalProperties: true, description: "Context bổ sung (field validation errors...)" }
        request_id: { type: string, description: "X-Request-ID để trace log" }
    Pagination:
      type: object
      properties:
        next_cursor: { type: string, nullable: true, description: "null nếu hết trang" }
        has_more:    { type: boolean }
        total:       { type: integer, description: "Tổng số records (optional, tính riêng nếu cần)" }
    User:
      type: object
      properties:
        id:    { type: string, format: uuid }
        email: { type: string }
        name:  { type: string }
        role:  { type: string, enum: [platform_admin, user] }
...
    Page:
      type: object
      properties:
        id:                   { type: string, format: uuid }
        site_id:              { type: string, format: uuid }
        current_version_id:   { type: string, format: uuid, nullable: true }
        slug:                 { type: string }
        status:               { type: string, enum: [draft, published] }
        seo_meta:             { type: object, additionalProperties: true }
        created_at:           { type: string, format: date-time }
        updated_at:           { type: string, format: date-time }
    PagedPages:
      allOf:
        - $ref: '#/components/schemas/Pagination'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/Page' }
...
    ContentTypeInput:
      type: object
      required: [name, field_schema]
      properties:
        name:             { type: string, example: Blog }
        requires_approval: { type: boolean, default: false }
        field_schema:
          type: array
          items:
            type: object
            properties:
              key:      { type: string }
              label:    { type: string }
              type:     { type: string, enum: [text, richtext, image, number, date, reference] }
              required: { type: boolean }
    ContentType:
      allOf:
        - $ref: '#/components/schemas/ContentTypeInput'
        - type: object
          properties:
            id:         { type: string, format: uuid }
            site_id:    { type: string, format: uuid }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
    ContentItem:
      type: object
      properties:
        id:               { type: string, format: uuid }
        content_type_id:  { type: string, format: uuid }
        site_id:          { type: string, format: uuid }
        fields:           { type: object, additionalProperties: true }
        status:           { type: string, enum: [draft, published] }
        created_at:       { type: string, format: date-time }
        updated_at:       { type: string, format: date-time }
    PagedContentItems:
      allOf:
        - $ref: '#/components/schemas/Pagination'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/ContentItem' }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
...